[rename] yaml config loader
